requirements.txt
numpy>=1.14.0
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
Pillow>=5.2.0
opencv-python>=4.5.5.64
tensorflow>=1.9.0
keras>=2.2.0
Flask>=1.1.2
mtcnn>=0.1.0
retina-face>=0.0.1
fire>=0.4.0
gunicorn>=20.1.0
Deprecated>=1.2.13
pip~=23.0.1
wheel~=0.40.0
Jinja2~=3.1.2
rsa~=4.9
pyasn1~=0.4.8
packaging~=23.0
scipy~=1.10.1
click~=8.1.3
gast~=0.4.0
six~=1.16.0
h5py~=3.8.0
pytz~=2023.3
requests~=2.28.2
colorama~=0.4.6
idna~=3.4
MarkupSafe~=2.1.2
Werkzeug~=2.2.3
itsdangerous~=2.1.2
wrapt~=1.14.1
setuptools~=67.6.1
mpmath~=1.3.0
sympy~=1.11.1
cachetools~=5.3.0
certifi~=2022.12.7
urllib3~=1.26.15
python-dateutil~=2.8.2
tifffile~=2023.4.12
Markdown~=3.4.3
oauthlib~=3.2.2
termcolor~=2.2.0
jax~=0.4.8
tensorboard~=2.12.2
flatbuffers~=23.3.3
astunparse~=1.6.3